requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.0
opencv-python~=4.11.0.86
python-nmap~=0.7.1
PyJWT~=2.10.1
Hypercorn~=0.17.3
Quart~=0.20.0
SQLAlchemy~=2.0.40
numpy~=2.2.4
aiosqlite>=0.18.0
celery==5.5.3
redis==6.2.0
aiosmtplib==4.0.1
email_validator==2.2.0
httpx==0.28.1
ffmpeg-python~=0.2.0
aiogram~=3.20.0.post0